fixed image positioning on iphone 6 and zooming in text backgrounds and publishing text size iophone 5

+++ VerbatmIOS/Verbatm/MediaUploader.m
@@ -79,7 +79,7 @@
 				NSLog(@"media upload progress: %ld out of %ld", (long)newProgressUnits, (long)self.mediaUploadProgress.totalUnitCount);
 			}
 		} success:^(NSURLSessionDataTask * _Nonnull task, NSData* responseData) {
-			[[PublishingProgressManager sharedInstance] mediaSavingProgressed:(self.mediaUploadProgress.totalUnitCount - (NSInteger)self.mediaUploadProgress.completedUnitCount)];
+			[[PublishingProgressManager sharedInstance] mediaSavingProgressed:((NSInteger)self.mediaUploadProgress.totalUnitCount - (NSInteger)self.mediaUploadProgress.completedUnitCount)];
 			[self.mediaUploadProgress setCompletedUnitCount: self.mediaUploadProgress.totalUnitCount];
 			NSLog(@"Image published!");
 			NSString *response = [[NSString alloc] initWithData:responseData encoding: NSUTF8StringEncoding];
